feat: enhance user registration with avatar upload and improve error handling

+++ Backend/utils/cloudinary.js
@@ -1,5 +1,8 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+
+dotenv.config();
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -19,7 +22,8 @@ const uploadOnCloudinary = async (localFilePath) => {
         fs.unlinkSync(localFilePath);
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+        console.error("Cloudinary Upload Error:", error);
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
         return null;
     }
 };
